requirements.txt
numpy
pandas
scikit-learn
xgboost
shap
nltk
pyyaml
torch
torchvision
transformers
accelerate
bitsandbytes
requests
